Updated examples
