+++ spec/examples/usual/dynamic_options/format/date/basic/example1_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe Usual::DynamicOptions::Format::Date::Example2 do
+RSpec.describe Usual::DynamicOptions::Format::Date::Basic::Example1 do
   describe ".call!" do
     subject(:perform) { described_class.call!(**attributes) }
 
@@ -14,7 +14,7 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Example2 do
 
     include_examples "check class info",
                      inputs: %i[started_on],
-                     internals: %i[started_on],
+                     internals: %i[],
                      outputs: %i[started_on]
 
     context "when the input arguments are valid" do
@@ -34,9 +34,8 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Example2 do
           it "returns expected error" do
             expect { perform }.to(
               raise_error(
-                ApplicationService::Exceptions::Internal,
-                "[Usual::DynamicOptions::Format::Date::Example2] Internal attribute `started_on` " \
-                "does not match `date` format"
+                ApplicationService::Exceptions::Input,
+                "[Usual::DynamicOptions::Format::Date::Basic::Example1] Input `started_on` does not match `date` format"
               )
             )
           end
@@ -66,7 +65,7 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Example2 do
 
     include_examples "check class info",
                      inputs: %i[started_on],
-                     internals: %i[started_on],
+                     internals: %i[],
                      outputs: %i[started_on]
 
     context "when the input arguments are valid" do
@@ -86,9 +85,8 @@ RSpec.describe Usual::DynamicOptions::Format::Date::Example2 do
           it "returns expected error" do
             expect { perform }.to(
               raise_error(
-                ApplicationService::Exceptions::Internal,
-                "[Usual::DynamicOptions::Format::Date::Example2] Internal attribute `started_on` " \
-                "does not match `date` format"
+                ApplicationService::Exceptions::Input,
+                "[Usual::DynamicOptions::Format::Date::Basic::Example1] Input `started_on` does not match `date` format"
               )
             )
           end
